+++ apps/3001-shop/next.config.js
@@ -4,10 +4,18 @@ const NextFederationPlugin = require('@module-federation/nextjs-mf');
 const {
   promiseTemplate,
 } = require('@module-federation/nextjs-mf/utils/build-utils');
+const {createDelegatedModule} = require("@module-federation/utilities");
 /**
  * @type {import('@nrwl/next/plugins/with-nx').WithNxOptions}
  **/
 const nextConfig = {
+  async redirects() {
+    return [{
+      source: '/',
+      destination: '/shop',
+      permanent: true
+    }];
+  },
   nx: {
     // Set this to true if you would like to to use SVGR
     // See: https://github.com/gregberge/svgr
@@ -22,25 +30,18 @@ const nextConfig = {
           name: 'shop',
           filename: 'static/chunks/remoteEntry.js',
           remotes: {
-
-  // shop: promiseTemplate('global@url', (resolve,reject) => {}),
-    home: promiseTemplate(
-    // can also be a string if it needs to be computed in scope
-    `(resolve, reject) => {
-                resolve("home_app@http://localhost:3000/_next/static/${
-      isServer ? 'ssr' : 'chunks'
-    }/remoteEntry.js");
-              }`,
-    (resolve,reject)=>{
-      console.log('runing other promise');
-      setTimeout(() => {
-        console.log('resolving promise');
-        resolve();
-      } , 1000);
-    }),
-            // home: `home_app@http://localhost:3000/_next/static/${
-            //   isServer ? 'ssr' : 'chunks'
-            // }/remoteEntry.js`,
+            // home: createDelegatedModule(require.resolve('./remote-delegate.js'), {
+            //   remote: `home_app@http://localhost:3000/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,
+            // }),
+            // shop: createDelegatedModule(require.resolve('./remote-delegate.js'), {
+            //   remote: `shop@http://localhost:3001/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,
+            // }),
+            // checkout: createDelegatedModule(require.resolve('./remote-delegate.js'), {
+            //   remote: `checkout@http://localhost:3002/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,
+            // }),
+            home: `home_app@http://localhost:3000/_next/static/${
+              isServer ? 'ssr' : 'chunks'
+            }/remoteEntry.js`,
             shop: `shop@http://localhost:3001/_next/static/${
               isServer ? 'ssr' : 'chunks'
             }/remoteEntry.js`,
@@ -59,6 +60,7 @@ const nextConfig = {
           },
           extraOptions: {
             exposePages: true,
+            automaticAsyncBoundary: true,
             enableImageLoaderFix: true,
             enableUrlLoaderFix: true,
             automaticPageStitching: true,