+++ apps/3002-checkout/next.config.js
@@ -1,11 +1,19 @@
 const {withNx} = require('@nrwl/next/plugins/with-nx');
 
 const NextFederationPlugin = require('@module-federation/nextjs-mf');
+const {createDelegatedModule} = require("@module-federation/utilities");
 
 /**
  * @type {import('@nrwl/next/plugins/with-nx').WithNxOptions}
  **/
 const nextConfig = {
+  async redirects() {
+    return [{
+      source: '/',
+      destination: '/checkout',
+      permanent: true
+    }];
+  },
   nx: {
     // Set this to true if you would like to to use SVGR
     // See: https://github.com/gregberge/svgr
@@ -19,9 +27,15 @@ const nextConfig = {
         name: 'checkout',
         filename: 'static/chunks/remoteEntry.js',
         remotes: {
-          home: `home_app@http://localhost:3000/_next/static/${
-            isServer ? 'ssr' : 'chunks'
-          }/remoteEntry.js`,
+          // home: createDelegatedModule(require.resolve('./remote-delegate.js'), {
+          //   remote: `home_app@http://localhost:3000/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,
+          // }),
+          // shop: createDelegatedModule(require.resolve('./remote-delegate.js'), {
+          //   remote: `shop@http://localhost:3001/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,
+          // }),
+          // checkout: createDelegatedModule(require.resolve('./remote-delegate.js'), {
+          //   remote: `checkout@http://localhost:3002/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,
+          // }),
           shop: `shop@http://localhost:3001/_next/static/${
             isServer ? 'ssr' : 'chunks'
           }/remoteEntry.js`,