+++ packages/node/src/plugins/LoadFileChunkLoadingRuntimeModule.ts
@@ -89,7 +89,8 @@ class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
     // so for example, if im in hostA and require(remoteb/module) --> console.log of name in runtime code will return remoteb
 
     const {remotes = {}, name} = this.options;
-
+    // for delegate modules, we need to avoid serializing internal modules, only register primitive configs in the runtime
+    // delegates have their own registration code, so we dont need to handle them here.
     const remotesByType: RemotesByType = Object.values(remotes).reduce((acc:RemotesByType, remote:string) => {
       if (remote.startsWith('promise ') || remote.startsWith('internal ') || remote.startsWith('external ')) {
         acc.functional.push(remote);
@@ -300,7 +301,7 @@ class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
                         /*
                       TODO: keying by global should be ok, but need to verify - need to deal with when user passes promise new promise() global will/should still exist - but can only be known at runtime
                     */
-                        this._getLogger(`'remotes keyed by global name'`, JSON.stringify(remotes)),
+                        this._getLogger(`'remotes keyed by global name'`, JSON.stringify(remotesByType.normal)),
                         this._getLogger(`'remote scope configs'`, 'global.__remote_scope__._config'),
 
                         this._getLogger(`'before remote scope'`),
@@ -352,7 +353,7 @@ class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
                         `loadScript(scriptUrl.toString(), function(err, content) {`,
                         Template.indent([
                           this._getLogger(`'load script callback fired'`),
-                          "if(err) {console.error('error loading remote chunk', scriptUrl.toString(),'got',content); return reject(err);}",
+                          "if(err) {console.error('error loading remote chunk', scriptUrl.toString(),'got',content,'with error', err); return reject(err);}",
                           'var chunk = {};',
                           'try {',
                           "require('vm').runInThisContext('(function(exports, require, __dirname, __filename) {' + content + '\\n})', filename)" +
@@ -362,7 +363,7 @@ class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
                           '}',
                           'installChunk(chunk);',
                         ]),
-                        '});',
+                        '}, chunkId);',
                       ]),
                       '}',
                     ]),