+++ packages/utilities/src/types/index.ts
@@ -57,8 +57,10 @@ export type ExternalsType = Required<
 type ModulePath = string;
 
 export type WebpackRemoteContainer = {
+  __initialized?: boolean;
   get(modulePath: ModulePath): () => any;
   init: (obj?: typeof __webpack_share_scopes__) => void;
+
 };
 
 export type AsyncContainer = Promise<WebpackRemoteContainer>;
@@ -85,3 +87,9 @@ export type Loader = Extract<RuleSetRule['use'], { loader?: string }>;
 export type EventTypes = 'loadStart' | 'loadComplete' | 'loadError';
 type NextRoute = string;
 export type PageMap = Record<NextRoute, ModulePath>;
+
+export type GetModuleOptions = {
+  modulePath: string
+  exportName?: string
+  remoteContainer: string | RemoteData
+}