+++ packages/node/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.10.3](https://github.com/module-federation/nextjs-mf/compare/node-0.10.2...node-0.10.3) (2023-01-31)
+
+### Dependency Updates
+
+* `utils` updated to version `1.2.0`
+
+
+## [0.10.2](https://github.com/module-federation/nextjs-mf/compare/node-0.10.1...node-0.10.2) (2023-01-21)
+
+### Dependency Updates
+
+* `utils` updated to version `1.1.2`
+
+
 ## [0.10.1](https://github.com/module-federation/nextjs-mf/compare/node-0.10.0...node-0.10.1) (2023-01-11)
 
 ### Dependency Updates